+++ test_classes.py
@@ -12,17 +12,17 @@ converter = Converter()
 # message = converter.history(limit=2, another="a")
 # print(message)
 
-message = converter.convert("length", "m yd 5")
-print(str(message))
+# message = converter.convert("length", "m 10")
+# print(str(message))
 
-# result = converter.manage_group("metro", "add metroo ", print_message=True)
+#converter.manage_group("metro", "remove")
+#print(f"Result: {result}")
+
+# result = converter.manage_type("temperature", "add new_type 1 1")
 # print(f"Result: {result}")
 
-# result = converter.manage_type("metro", "metru add 0.00001", print_message=True)
-# print(f"Result: {result}")
+result = converter.change_base("length", "yard")
+print(f"Result: {result}")
 
-# result = converter.change_base("metro", "metroo", print_message=True)
-# print(f"Result: {result}")
-
-# result = converter.aliases("metro", "metroo remove mo", print_message=True)
+# result = converter.aliases("length", "remove meters mtr", print_message=True)
 # print(f"Result: {result}")